refactor(arch): Phase 3 — PR #483 리뷰 수정
- B1: byokGate.ts → shared/lib/ 이동 (cross-entity dependency 해소)
- S1: submitOverallAnalysisAction.test.ts barrel requireActual로 전환
