+++ eslint.config.mjs
@@ -125,8 +125,14 @@ const eslintConfig = defineConfig([
                         {
                             from: 'shared',
                             // 마이그레이션 중 임시 허용: shared로 이동한 파일이 아직 legacy-domain 타입/상수를 참조.
-                            // Phase 3 (entities 마이그레이션) 완료 시 legacy-domain 제거.
-                            allow: ['shared', 'legacy-domain'],
+                            // Phase 3: byokGate가 shared/lib/로 이동하면서 entities(api-key, user)와
+                            // legacy-infra(tier)를 참조. Phase 5 (features 분리) 완료 시 entities/legacy-infra 제거.
+                            allow: [
+                                'shared',
+                                'entities',
+                                'legacy-domain',
+                                'legacy-infra',
+                            ],
                         },
                         // 옛 layer 간 의존: 현재 코드 그대로 허용
                         {
