+++ src/entities/analysis/actions/submitOverallAnalysisAction.ts
@@ -20,7 +20,7 @@ import {
 } from '@/entities/news-article';
 import { getNextEarningsReport } from '@/entities/earnings-report';
 import { getCurrentUser } from '@/infrastructure/auth/getCurrentUser';
-import { resolveTierAndByok, buildGateError } from '../lib/byokGate';
+import { resolveTierAndByok, buildGateError } from '@/shared/lib/byokGate';
 import { isBot } from '@/shared/api/isBot';
 import { fetchOptionsSnapshot } from '@/infrastructure/options/optionsDataCache';
 import {
