+++ src/entities/analysis/index.ts
@@ -9,11 +9,5 @@ export {
     releaseReanalyzeCooldown,
     getReanalyzeCooldownMs,
 } from './lib/reanalyzeCooldown';
-export {
-    resolveTierAndByok,
-    buildGateError,
-    isKnownModelId,
-    type ByokOutcome,
-} from './lib/byokGate';
 
 // actions are imported from @/entities/analysis/actions
