+++ src/shared/lib/byokGate.ts
